refactor(web): the admin dashboard as its screens
The 4,300-line file becomes twenty modules under `src/admin/` and a
150-line shell: the addresses and views in `routing.ts`, the number and
date wording in `format.ts`, one file per chart, the accounts table with
its sort engine, the page chrome, and one file per screen. Each
`Page`/`Screen` pair merges — with the read hook there was nothing left
to split but fifteen props handed down a level — and the sidebar's fold
geometry folds into the one component that draws it.

`PRIVACY.md` now says the admin page exists and what a maintainer can see
on it, which it never did.
