+++ apps/web/src/admin/auth-client.ts
@@ -1,0 +1,3 @@
+import { createAuthClient } from "better-auth/react";
+
+export const authClient = createAuthClient();
