+++ apps/web/src/admin/chrome/cards.tsx
@@ -1,0 +1,136 @@
+import { useState } from "react";
+import { GitHubMark, GoogleMark } from "../../account-marks";
+import { AUTH_BUTTON } from "../../auth-surface";
+import { LukeMark } from "../../SiteChrome";
+import {
+  SOCIAL_PROVIDER,
+  SOCIAL_PROVIDER_LABEL,
+  type SocialProvider,
+} from "../../sign-in-provider";
+import { authClient } from "../auth-client";
+import { SIGN_IN_CHOSEN } from "../prefs";
+import { PLAIN_BUTTON } from "./controls";
+
+export function SignInCard(): React.JSX.Element {
+  const [pending, setPending] = useState<SocialProvider>();
+  const [failed, setFailed] = useState(false);
+
+  const begin = async (provider: SocialProvider) => {
+    if (pending) return;
+    setPending(provider);
+    setFailed(false);
+    const result = await authClient.signIn.social({
+      provider,
+      callbackURL: window.location.pathname,
+    });
+    if (result.error) {
+      setPending(undefined);
+      setFailed(true);
+      return;
+    }
+    SIGN_IN_CHOSEN.write(true);
+  };
+
+  return (
+    <main className="grid min-h-screen place-items-center p-6">
+      <section className="w-full max-w-[390px] rounded-lg border border-border bg-card px-8 py-12 text-center shadow-[0_24px_80px_rgba(0,0,0,0.42)]">
+        <div className="inline-flex w-13" aria-hidden="true">
+          <LukeMark className="h-auto w-full" />
+        </div>
+        <h1 className="mt-4 mb-2 text-[1.75rem] leading-[1.15] font-semibold">Luke admin</h1>
+        <p className="m-0 text-muted-foreground">
+          Sign in with the account this dashboard is restricted to.
+        </p>
+        <div className="mt-8 mb-4 grid gap-3">
+          <button
+            type="button"
+            className={`${AUTH_BUTTON} inline-flex items-center justify-center gap-2.5`}
+            disabled={pending !== undefined}
+            onClick={() => void begin(SOCIAL_PROVIDER.GOOGLE)}
+          >
+            <GoogleMark className="size-[15px] shrink-0" />
+            {pending === SOCIAL_PROVIDER.GOOGLE
+              ? "Opening…"
+              : `Continue with ${SOCIAL_PROVIDER_LABEL[SOCIAL_PROVIDER.GOOGLE]}`}
+          </button>
+          <button
+            type="button"
+            className={`${AUTH_BUTTON} inline-flex items-center justify-center gap-2.5`}
+            disabled={pending !== undefined}
+            onClick={() => void begin(SOCIAL_PROVIDER.GITHUB)}
+          >
+            <GitHubMark className="size-[15px] shrink-0" />
+            {pending === SOCIAL_PROVIDER.GITHUB
+              ? "Opening…"
+              : `Continue with ${SOCIAL_PROVIDER_LABEL[SOCIAL_PROVIDER.GITHUB]}`}
+          </button>
+        </div>
+        {failed ? <p className="m-0 text-attention">Sign-in could not start. Try again.</p> : null}
+      </section>
+    </main>
+  );
+}
+
+export function Centered({
+  title,
+  children,
+}: {
+  title: string;
+  children: React.ReactNode;
+}): React.JSX.Element {
+  return (
+    <main className="grid min-h-screen place-items-center p-6">
+      <section className="w-full max-w-[420px] rounded-lg border border-border bg-card px-8 py-12 text-center">
+        <div className="mx-auto inline-flex w-11" aria-hidden="true">
+          <LukeMark className="h-auto w-full" />
+        </div>
+        <h1 className="mt-4 mb-2 text-2xl font-semibold">{title}</h1>
+        <div className="text-muted-foreground">{children}</div>
+      </section>
+    </main>
+  );
+}
+
+/** The gate's one non-admin refusal, worded the same on both views. */
+export function ForbiddenCard({
+  email,
+  onSignOut,
+}: {
+  email: string | undefined;
+  onSignOut: () => void;
+}): React.JSX.Element {
+  return (
+    <Centered title="Not authorized">
+      You are signed in{email ? ` as ${email}` : ""}, but this account does not have the admin role.
+      Admin access is the <code className="font-mono">admin</code> role on your account, set
+      directly in the database.
+      <div className="mt-6">
+        <button type="button" className={PLAIN_BUTTON} onClick={onSignOut}>
+          Sign out
+        </button>
+      </div>
+    </Centered>
+  );
+}
+
+/** The failed read's own card, with the retry every screen offers on it. */
+export function AdminErrorCard({
+  detail,
+  refreshing,
+  onRetry,
+}: {
+  detail: string;
+  refreshing: boolean;
+  onRetry: () => void;
+}): React.JSX.Element {
+  return (
+    <Centered title="Could not load">
+      {detail}
+      <div className="mt-6">
+        <button type="button" className={PLAIN_BUTTON} onClick={onRetry} disabled={refreshing}>
+          {refreshing ? "Trying…" : "Try again"}
+        </button>
+      </div>
+    </Centered>
+  );
+}
