+++ apps/web/src/admin/format.ts
@@ -1,0 +1,78 @@
+import type { AdminTrend } from "../../server/admin/admin-metrics";
+import { SOCIAL_PROVIDER, SOCIAL_PROVIDER_LABEL } from "../sign-in-provider";
+
+const numberFormat = new Intl.NumberFormat("en-US");
+
+export function formatNumber(value: number): string {
+  return numberFormat.format(value);
+}
+
+/** An instant drawn as its UTC date alone, e.g. "Aug 3, 2026". */
+export function formatDate(epochMs: number): string {
+  return new Date(epochMs).toLocaleDateString("en-US", {
+    dateStyle: "medium",
+    timeZone: "UTC",
+  });
+}
+
+/** A day key drawn as a short axis tick, e.g. "Aug 21". */
+export function formatDayTick(day: string): string {
+  return new Date(`${day}T00:00:00.000Z`).toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+    timeZone: "UTC",
+  });
+}
+
+/** A day key drawn as the day page's own masthead, e.g. "August 21, 2026". */
+export function formatDayHeading(day: string): string {
+  return new Date(`${day}T00:00:00.000Z`).toLocaleDateString("en-US", {
+    dateStyle: "long",
+    timeZone: "UTC",
+  });
+}
+
+/**
+ * How faded today's bar draws. The series' last day is still filling, so a
+ * bar rendered like a complete day's always reads as a dip; the fade and the
+ * tooltip's "(so far today)" say the day is partial instead.
+ */
+export const PARTIAL_DAY_OPACITY = 0.45;
+
+/** A tooltip label for one bar's day, saying so when that day is still filling. */
+export function formatTooltipDay(day: string, partialDay: string | undefined): string {
+  return day === partialDay ? `${formatDayTick(day)} (so far today)` : formatDayTick(day);
+}
+
+const TREND_TONE = {
+  UP: "text-complete",
+  DOWN: "text-attention",
+  FLAT: "text-muted-foreground",
+} as const;
+
+export function trendTone(trend: AdminTrend): string {
+  if (trend.recent > trend.prior) return TREND_TONE.UP;
+  if (trend.recent < trend.prior) return TREND_TONE.DOWN;
+  return TREND_TONE.FLAT;
+}
+
+/**
+ * How the run moved against the one before it, as a percentage where there is
+ * one to state. A prior run of zero has none — every rise from nothing is
+ * infinite — so the move stands as its own count rather than a figure that
+ * reads as precision it does not have.
+ */
+export function formatTrendMove(trend: AdminTrend): string {
+  const delta = trend.recent - trend.prior;
+  if (delta === 0) return "flat";
+  const sign = delta > 0 ? "+" : "−";
+  if (trend.prior === 0) return `${sign}${formatNumber(Math.abs(delta))}`;
+  return `${sign}${Math.abs(Math.round((delta / trend.prior) * 100))}%`;
+}
+
+/** A linked provider's row value drawn as its label where the page knows one. */
+export function signInMethodLabel(providerId: string): string {
+  if (providerId === SOCIAL_PROVIDER.GITHUB) return SOCIAL_PROVIDER_LABEL[SOCIAL_PROVIDER.GITHUB];
+  if (providerId === SOCIAL_PROVIDER.GOOGLE) return SOCIAL_PROVIDER_LABEL[SOCIAL_PROVIDER.GOOGLE];
+  return providerId;
+}
